Prepare new pull command
