+++ caldir-cli/src/commands/mod.rs
@@ -1,6 +1,6 @@
 // pub mod auth;
 // pub mod new;
-// pub mod pull;
+pub mod pull;
 // pub mod push;
 pub mod status;
